Move notifications models to shared
