--- apps/dashboard/src/entities/notifications/model/NotificationToPush.ts
@@ -1,3 +1,0 @@
-import type { Notification } from '@/src/entities/notifications/model/Notification'
-
-export type NotificationToPush = Omit<Notification, 'createdAt'>
