+++ apps/dashboard/src/features/algorithm/ui/MultipleBotsAlgorithmCard.tsx
@@ -3,7 +3,7 @@ import AlgorithmCardFrame from '@/src/entities/algorithm/ui/AlgorithmCardFrame'
 import RunAlgorithmModal from '@/src/entities/algorithm/ui/RunAlgorithmModal'
 import { trpc } from '@/src/shared/api/trpc'
 import { useAppDispatch } from '@/src/shared/model/hooks'
-import { pushNotification } from '@/src/entities/notifications/model/notificationsSlice'
+import { pushNotification } from '@/src/shared/model/notificationsSlice'
 
 export interface MultipleBotsAlgorithmCard {
   algorithm: Algorithm
