+++ apps/dashboard/src/features/notifications/NotificationsQueue.tsx
@@ -2,8 +2,8 @@ import { memo, useCallback } from 'react'
 
 import NotificationCard from '@/src/entities/notifications/ui/NotificationCard'
 import { useAppSelector, useAppDispatch } from '@/src/shared/model/hooks'
-import { removeNotification } from '@/src/entities/notifications/model/notificationsSlice'
-import { Notification } from '@/src/entities/notifications/model/Notification'
+import { removeNotification } from '@/src/shared/model/notificationsSlice'
+import { Notification } from '@/src/shared/model/Notification'
 
 const NotificationCardMemo = memo(NotificationCard)
