+++ apps/dashboard/src/shared/model/NotificationToPush.ts
@@ -1,0 +1,3 @@
+import type { Notification } from '@/src/shared/model/Notification'
+
+export type NotificationToPush = Omit<Notification, 'createdAt'>
