+++ apps/dashboard/src/widgets/algorithm/AlgorithmHeaderDescriptor.tsx
@@ -2,7 +2,7 @@ import type { Algorithm } from '@tradeb0t/core'
 import InputsDescriptor from '@/src/entities/algorithm/ui/InputsDescriptor'
 import { trpc } from '@/src/shared/api/trpc'
 import { useAppDispatch } from '@/src/shared/model/hooks'
-import { pushNotification } from '@/src/entities/notifications/model/notificationsSlice'
+import { pushNotification } from '@/src/shared/model/notificationsSlice'
 import { algorithmRunSuccessNotification } from '@/src/entities/algorithm/notifications/algorithmRunSuccessNotification'
 import { algorithmRunErrorNotification } from '@/src/entities/algorithm/notifications/algorithmRunErrorNotification'
 import RunAlgorithmForm from '@/src/entities/algorithm/ui/RunAlgorithmForm'
